backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
redis==5.0.8
rq==1.16.1
stripe==10.5.0
alembic==1.14.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
